refactor: 공통 로직 상속으로 변경
